#784: Move ui component to common folder and rename
